+++ src/main/java/org/board/exercise_board/user/controller/UserController.java
@@ -3,12 +3,12 @@ package org.board.exercise_board.user.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.board.exercise_board.user.application.SignInApplication;
 import org.board.exercise_board.user.application.SignUpApplication;
 import org.board.exercise_board.user.domain.Dto.UserDto;
 import org.board.exercise_board.user.domain.Form.SignInForm;
 import org.board.exercise_board.user.domain.Form.SignUpForm;
 import org.board.exercise_board.user.domain.model.JwtToken;
-import org.board.exercise_board.user.service.UserService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,8 +22,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/user")
 public class UserController {
 
-  private final UserService userService;
   private final SignUpApplication signUpApplication;
+  private final SignInApplication signInApplication;
 
   @Operation(summary = "회원가입")
   @PostMapping("/signup")
@@ -35,13 +35,13 @@ public class UserController {
   @Operation(summary = "이메일 인증")
   @GetMapping("/verified/{tokenId}")
   public String verifyEmail(@PathVariable("tokenId") String tokenId) {
-    userService.verifyEmail(tokenId);
+    signUpApplication.verifyEmail(tokenId);
     return "인증이 완료되었습니다.";
   }
 
   @Operation(summary = "로그인")
   @PostMapping("/signin")
   public ResponseEntity<JwtToken> signin(@Valid @RequestBody SignInForm signInForm) {
-    return ResponseEntity.ok(userService.signin(signInForm));
+    return ResponseEntity.ok(signInApplication.signIn(signInForm));
   }
 }
