+++ src/main/java/org/board/exercise_board/user/service/UserService.java
@@ -25,7 +25,6 @@ public class UserService implements UserDetailsService {
 
   private final UserRepository userRepository;
   private final JwtTokenProvider jwtTokenProvider;
-  private final EmailService emailService;
 
   private final PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 
@@ -63,23 +62,6 @@ public class UserService implements UserDetailsService {
     return token;
   }
 
-  public void verifyEmail(String tokenId) {
-
-    EmailToken emailToken = emailService.findToken(tokenId)
-            .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
-
-    // 이미 인증완료되었다면 Exception 발생
-    if(emailToken.getUser().getVerifiedStatus()) {
-      throw new CustomException(ErrorCode.ALREADY_VERIFIED);
-    }
-
-    // 만료 시간이 지났다면 Exception 발생
-    if(emailService.verifyExpirationDateTime(emailToken)) {
-      emailService.updateVerifyStatus(emailToken);
-    } else {
-      throw new CustomException(ErrorCode.EXPIRATION_TIME_IS_OVER);
-    }
-  }
 
   public boolean isExistEmail(String email) {
     return userRepository.existsByEmail(email);
